requirements.txt
asgiref                  
bcrypt                   
cffi                     
click                    
colorama                 
cryptography             
dj-database-url          
Django                   
django-core              
django-heroku            
django-managepy-anywhere 
Flask                    
flask-marshmallow        
Flask-SQLAlchemy         
greenlet                 
gunicorn                 
itsdangerous             
Jinja2                   
jsonify                  
manage.py                
MarkupSafe               
marshmallow              
marshmallow-sqlalchemy   
mysqlclient             
packaging                
pip                      
psycopg2                 
pycparser                
PyJWT                    
pymssql                  
PyMySQL                  
pyparsing                
setuptools               
six                      
SQLAlchemy               
sqlparse                 
tzdata                   
Werkzeug                 
whitenoise               